Race and Class APIs
