+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "marked": "^14.1.2"
+  }
+}
